# 0406-0305-scenario-simulator-snapshot/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Scenario Simulator
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Confidence = (typeof Confidence)[keyof typeof Confidence];

export const Confidence = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export type StepKey = (typeof StepKey)[keyof typeof StepKey];

export const StepKey = {
  brief: "brief",
  stakeholder: "stakeholder",
  interview: "interview",
  evidence: "evidence",
  define: "define",
  submit: "submit",
} as const;

/**
 * Map of step key to milliseconds accumulated on that step.
 */
export type StepTimingsTotals = { [key: string]: number };

/**
 * Cumulative time spent on each step plus the active step's start.
 */
export interface StepTimings {
  /** Map of step key to milliseconds accumulated on that step. */
  totals: StepTimingsTotals;
  currentStep: StepKey | null;
  currentStepStartedAt: string | null;
}

export type AnswerRecordSelected =
  (typeof AnswerRecordSelected)[keyof typeof AnswerRecordSelected];

export const AnswerRecordSelected = {
  A: "A",
  B: "B",
  C: "C",
} as const;

export interface AnswerRecord {
  questionId: string;
  selected: AnswerRecordSelected;
}

export interface Workshop {
  id: string;
  code: string;
  label: string;
  createdAt: string;
}

export interface CreateWorkshopInput {
  /** @minLength 1 */
  label: string;
  /**
   * Optional client-supplied code (4-6 chars, alphabet A-Z2-9, case-insensitive). Auto-generated when omitted.
   * @minLength 4
   * @maxLength 6
   * @pattern ^[A-Za-z2-9]+$
   */
  code?: string | null;
}

export interface Session {
  id: string;
  workshopId: string;
  workshopCode: string;
  teamName: string;
  currentScreen: string;
  selectedStakeholder: string | null;
  selectedEvidenceSource: string | null;
  answers: AnswerRecord[];
  problemStatement: string;
  confidence: Confidence | null;
  assumption: string;
  flaggedForDebrief: boolean;
  stepTimings: StepTimings;
  submittedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ModeratorNote {
  id: string;
  sessionId: string;
  message: string;
  templateId: string | null;
  createdAt: string;
  dismissedAt: string | null;
}

export type AccessRequestStatus =
  (typeof AccessRequestStatus)[keyof typeof AccessRequestStatus];

export const AccessRequestStatus = {
  pending: "pending",
  granted: "granted",
  declined: "declined",
} as const;

export interface AccessRequest {
  id: string;
  sessionId: string;
  status: AccessRequestStatus;
  createdAt: string;
  respondedAt: string | null;
}

export interface SetFlagInput {
  flagged: boolean;
}

export interface CreateNoteInput {
  /** @minLength 1 */
  message: string;
  templateId?: string | null;
}

export interface BroadcastNoteInput {
  /** @minLength 1 */
  message: string;
  templateId?: string | null;
  /** Restrict the broadcast to teams in this workshop. Required for multi-workshop deployments. */
  workshopCode: string;
}

export type RespondAccessInputStatus =
  (typeof RespondAccessInputStatus)[keyof typeof RespondAccessInputStatus];

export const RespondAccessInputStatus = {
  granted: "granted",
  declined: "declined",
} as const;

export interface RespondAccessInput {
  status: RespondAccessInputStatus;
}

export interface CreateSessionInput {
  /** @minLength 1 */
  workshopCode: string;
  /** @minLength 1 */
  teamName: string;
}

export interface UpdateSessionInput {
  currentScreen?: string;
  selectedStakeholder?: string | null;
  selectedEvidenceSource?: string | null;
  answers?: AnswerRecord[];
  problemStatement?: string;
  confidence?: Confidence | null;
  assumption?: string;
}

export interface SubmitSessionInput {
  /** @minLength 1 */
  problemStatement: string;
  confidence: Confidence;
  assumption: string;
  selectedStakeholder: string;
  selectedEvidenceSource: string;
}

export type ListSessionsParams = {
  /**
   * Optional filter to limit sessions to a single workshop
   */
  workshopCode?: string;
};

export type ListSubmissionsParams = {
  /**
   * Optional filter to limit submissions to a single workshop
   */
  workshopCode?: string;
};