# 0406-0305-scenario-simulator-snapshot/lib/api-zod/src/generated/types/accessRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Scenario Simulator
 * OpenAPI spec version: 0.1.0
 */
import type { AccessRequestStatus } from "./accessRequestStatus";

export interface AccessRequest {
  id: string;
  sessionId: string;
  status: AccessRequestStatus;
  createdAt: Date;
  respondedAt: Date | null;
}