# 0406-0305-scenario-simulator-snapshot/lib/api-zod/src/generated/types/createWorkshopInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Scenario Simulator
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWorkshopInput {
  /** @minLength 1 */
  label: string;
  /**
   * Optional client-supplied code (4-6 chars, alphabet A-Z2-9, case-insensitive). Auto-generated when omitted.
   * @minLength 4
   * @maxLength 6
   * @pattern ^[A-Za-z2-9]+$
   */
  code?: string | null;
}