# 0406-0305-scenario-simulator-snapshot/lib/api-zod/src/generated/types/respondAccessInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Scenario Simulator
 * OpenAPI spec version: 0.1.0
 */
import type { RespondAccessInputStatus } from "./respondAccessInputStatus";

export interface RespondAccessInput {
  status: RespondAccessInputStatus;
}